pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-core-py"
version = "0.1.0"
description = "A clean Python runtime for multi-provider LLM apps and agent workflows"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Shlok Mundhra" }
]
keywords = [
    "llm",
    "ai",
    "agent",
    "openai",
    "anthropic",
    "gemini",
    "deepseek",
    "multi-provider",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.18.0"]
gemini = ["google-generativeai>=0.4.0"]
deepseek = ["openai>=1.0.0"]
all = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.4.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
agent = "agent.cli.main:main"

[project.urls]
Homepage = "https://github.com/THE-CRED/Agent-core"
Documentation = "https://agent-core.readthedocs.io"
Repository = "https://github.com/THE-CRED/Agent-core"

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=agent --cov-report=term-missing"
